Revert "Merge pull request #13 from CosmosOS/master"
This reverts commit e2b71f0e5cd46fde5748b753774c3c6164c8acd2, reversing
changes made to 50be0f78824a9cb1e730ccebbfff9ac8758d0051.

+++ source/Cosmos.System2/FileSystem/FAT/FatStream.cs
@@ -201,7 +201,9 @@ namespace Cosmos.System.FileSystem.FAT
                     xReadSize = xCount;
                 }
 
-                Array.Copy(xCluster, xPosInCluster, aBuffer, xOffset, xReadSize);
+                //TODO: .Net Core
+                //Array.Copy(xCluster, xPosInCluster, aBuffer, xOffset, xReadSize);
+                Array.Copy(xCluster, (int)xPosInCluster, aBuffer, (int)xOffset, (int)xReadSize);
 
                 xOffset += xReadSize;
                 xCount -= xReadSize;
